+++ analyzers/tests/SonarAnalyzer.Test/Extensions/ITypeSymbolExtensionsTest.cs
@@ -37,7 +37,7 @@ public class ITypeSymbolExtensionsTest
     public void Compile()
     {
         using var workspace = new AdhocWorkspace();
-        var document = workspace.CurrentSolution.AddProject("foo", "foo.dll", LanguageNames.CSharp).AddDocument("test", SymbolHelperTest.TestInput);
+        var document = workspace.CurrentSolution.AddProject("foo", "foo.dll", LanguageNames.CSharp).AddDocument("test", ISymbolExtensionsTest.TestInput);
         var compilation = document.Project.GetCompilationAsync().Result;
         var tree = compilation.SyntaxTrees.First();
         semanticModel = compilation.GetSemanticModel(tree);
@@ -522,6 +522,77 @@ public class ITypeSymbolExtensionsTest
             }
             """);
         fieldSymbol.Type.IsEnum().Should().BeFalse();
+    }
+
+    [TestMethod]
+    public void GetSelfAndBaseTypes_WhenSymbolIsNull_ReturnsEmpty() =>
+        ((ITypeSymbol)null).GetSelfAndBaseTypes().Should().BeEmpty();
+
+    [DataTestMethod]
+    [DataRow("BaseClass<int>         ", "InheritedAttribute", "DerivedInheritedAttribute", "DerivedNotInheritedAttribute", "UnannotatedAttribute", "NotInheritedAttribute")]
+    [DataRow("DerivedOpenGeneric<int>", "InheritedAttribute", "DerivedInheritedAttribute", "DerivedNotInheritedAttribute", "UnannotatedAttribute")]
+    [DataRow("DerivedClosedGeneric   ", "InheritedAttribute", "DerivedInheritedAttribute", "DerivedNotInheritedAttribute", "UnannotatedAttribute")]
+    [DataRow("Implement              ")]
+    public void GetAttributesWithInherited_TypeSymbol(string className, params string[] expectedAttributes)
+    {
+        className = className.TrimEnd();
+        var code = $$"""
+                using System;
+
+                [AttributeUsage(AttributeTargets.All, Inherited = true)]
+                public class InheritedAttribute : Attribute { }
+
+                [AttributeUsage(AttributeTargets.All, Inherited = false)]
+                public class NotInheritedAttribute : Attribute { }
+
+                public class DerivedInheritedAttribute: InheritedAttribute { }
+
+                public class DerivedNotInheritedAttribute: NotInheritedAttribute { }
+
+                public class UnannotatedAttribute : Attribute { }
+
+                [Inherited]
+                [DerivedInherited]
+                [NotInherited]
+                [DerivedNotInherited]
+                [Unannotated]
+                public class BaseClass<T1> { }
+
+                [Inherited]
+                [DerivedInherited]
+                [NotInherited]
+                [DerivedNotInherited]
+                [Unannotated]
+                public interface IInterface { }
+
+                public class DerivedOpenGeneric<T1>: BaseClass<T1> { }
+
+                public class DerivedClosedGeneric: BaseClass<int> { }
+
+                public class Implement: IInterface { }
+
+                public class Program
+                {
+                    public static void Main()
+                    {
+                        new {{className}}();
+                    }
+                }
+                """;
+        var compiler = new SnippetCompiler(code);
+        var objectCreation = compiler.GetNodes<ObjectCreationExpressionSyntax>().Should().ContainSingle().Subject;
+        if (compiler.GetSymbol<IMethodSymbol>(objectCreation) is { MethodKind: MethodKind.Constructor, ReceiverType: { } receiver })
+        {
+            var actual = receiver.GetAttributesWithInherited().Select(x => x.AttributeClass.Name).ToList();
+            actual.Should().BeEquivalentTo(expectedAttributes);
+        }
+        else
+        {
+            Assert.Fail("Constructor could not be found.");
+        }
+        // GetAttributesWithInherited should behave like MemberInfo.GetCustomAttributes from runtime reflection:
+        var type = compiler.EmitAssembly().GetType(className.Replace("<int>", "`1"), throwOnError: true);
+        type.GetCustomAttributes(inherit: true).Select(x => x.GetType().Name).Should().BeEquivalentTo(expectedAttributes);
     }
 
     private static IFieldSymbol FirstFieldSymbolFromCode(string code)
